refactor: file cleanup
